Added ability to place towers
Added a "Tile" script and a "BuildManager" script. the "Tile" script checks to see if there is a tower on a given tile and if not allows one to be placed. "Tile" gets the tower to be built from the "BuildManager". This code will need to be expanded on once more towers are created. Also removed some old commented out code from the "Enemy" script