Tower Placement Rework is complete
tower placement is now freeform as opposed to tile based. Tower selection/deselection, selling and upgrading is now adjusted to work with the new tower placement. upgrade buttons and tower purchase buttons now are not interactable if the player does not have enough money. wile placing tower the tower will change color if it is not able to be placed (red) or is able to be placed (blue).